Remove intermediary yaml format
