--- mapping/yaml_formats.go
@@ -1,13 +1,0 @@
-package mapping
-
-type yamlMapping struct {
-	Static []mapping
-	Proxy  []mapping
-}
-
-type mapping struct {
-	From   string
-	Inject Injection
-	Regexp string
-	To     string
-}
